Remove boostrap-specific component

+++ airframe-rx-widget/src/main/scala/wvlet/airframe/rx/html/widget/auth/GoogleAuth.scala
@@ -92,7 +92,7 @@ class GoogleAuth(config: GoogleAuthConfig) extends LogSupport {
       )
 
       // Refresh auth token
-      timers.setInterval(config.tokenRefreshIntervalMillis) {
+      timers.setInterval(config.tokenRefreshIntervalMillis.toDouble) {
         refreshAuth
       }
     }
